feat(#19) :: interface 타입스로 빼기
