+++ src/hooks/useSurvey.ts
@@ -1,13 +1,13 @@
 import { getSurvey } from "@/services/survey";
-import { useQuery } from "@tanstack/react-query";
+import { UseQueryResult, useQuery } from "@tanstack/react-query";
 import { Survey } from "@/services/survey";
 
-export function useGetSurvey() {
-    return useQuery<Survey>({
+export function useGetSurvey(): UseQueryResult<Survey, Error> {
+    return useQuery<Survey, Error>({
         queryKey: ["getSurvey"],
         queryFn: async () => {
             const res = await getSurvey();
-            return res || [];
+            return res;
         },
         enabled: false,
     })
