+++ src/services/auth.ts
@@ -1,13 +1,7 @@
 import customAxois from '@/lib/customAxios';
 
-export interface AuthResponse {
-  access_token: string;
-  user: {
-    id: number;
-    name: string;
-    profile: string;
-  };
-}
+
+import { AuthResponse } from '@/types/auth';
 
 export async function kakaoLogin(code: string): Promise<AuthResponse> {
   const res = await customAxois.get<AuthResponse>("/auth/kakao/callback", {
