+++ src/services/survey.ts
@@ -1,11 +1,14 @@
 import customAxois from '@/lib/customAxios';
 
-export interface Survey {
-    data: [];
-    count: number;
+
+import { Survey } from '@/types/survey';
+
+export async function getSurvey(): Promise<Survey> {
+  const res = await customAxois.get<Survey>("/survey");
+  return res.data;
 }
 
-export async function getSurvey<Survey>() {
-  const res = await customAxois.get<Survey>("/survey");
+export async function createSurvey() {
+  const res = await customAxois.post<Survey>("/survey");
   return res.data;
 }
