+++ src/types/survey.ts
@@ -1,0 +1,4 @@
+export interface Survey {
+    data: [];
+    count: number;
+}
